Configure the intellij checkstyle plugin to use the correct checkstyle version
Currently, we do not set the CheckStyle-IDEA version, so it defaults to the latest, 8.1. There was a breaking change introduced in 8.1 that makes the baseline checkstyle files invalid (https://github.com/checkstyle/checkstyle/issues/4714). This means you get an error message and the checkstyle plugin no longer works in IDEA. It's also sadness to have the IDE and gradle running different checkstyle versions.