[improvement] Enforce lang3 Validate calls with no parameters and constant strings use logsafe (#521)
## Before this PR
There is a lower degree of standardization in the use of basic calls to verify conditions, specifically as it pertains to the use of commons-lang3 Validate calls.

## After this PR
Simple use cases of commons-lang3 Validate.isTrue, Validate.notNull, Validate.validState are now required to be changed to use logsafe variants. This will help to standardize calls where it is clear there can be a direct one-for-one replacement made with logsafe calls.