[feature] New 'com.palantir.baseline-reproducibility' plugin (#539)
## Before this PR

I discovered that in most of our projects, running `./gradlew jar` twice on a completely clean checkout results in jars with different hashes. This defeats any kind of fancy gradle caching we might want to enable in the future.

## After this PR

Projects can `apply plugin: 'com.palantir.baseline-reproducibility'` to opt-in to more sensible behaviour.