Fix incorrect argument selection in ValidateConstantMessage and expand test functionality. (#546)
## Before this PR
The ValidateConstantMessage check would flag on the last argument that is a string rather than on the message string. Additionally, the grouping of test cases under the single positive test does not correctly test cases after the first one.

## After this PR
The ValidateConstantMessage check now knows the argument index of all message argument in all calls so it can be correctly identified and checked. Test cases are now all completely separated in their own tests.