Disallow javafx imports with checkstyle (#569)
<!-- PR title should start with '[fix]', '[improvement]' or '[break]' if this PR would cause a patch, minor or major SemVer bump. Omit the prefix if this PR doesn't warrant a standalone release. -->

## Before this PR
Checkstyle allows javafx imports. javafx comes with Oracle JVMs, but is not included in some others we use internally. My team's internal project has a checkstyle customization that does this (we got broken by a javafx `Pair` import) but we'd like to be able to switch to the standard checkstyle and avoid another service hitting the same issue.

## After this PR
Disallow javafx imports. There's usually a non-javafx alternative.

CC @schlosna