[break] Enable running of unique class check on multiple configurations (#583)
<!-- PR title should start with '[fix]', '[improvement]' or '[break]' if this PR would cause a patch, minor or major SemVer bump. Omit the prefix if this PR doesn't warrant a standalone release. -->

## Before this PR
<!-- Describe the problem you encountered with the current state of the world (or link to an issue) and why it's important to fix now. -->
For e.g. spark docker builds we need to be able to check the class uniqueness on two different configurations.

## After this PR
<!-- Describe at a high-level why this approach is better. -->
An optional extension is added to add different configurations to check. If nothing is specified, the build will continue function as normal.

<!-- Reference any existing GitHub issues, e.g. 'fixes #000' or 'relevant to #000' -->

## Break context

This changes gradle task names (`checkClassUniqueness` => `checkRuntimeClassUniqueness`). I can set up an dummy task to keep back-compat if preferable.