Don't cache test tasks (#694)
Don't cache test tasks in the build cache

This is to remediate the situation where the inputs of test tasks aren't properly configured. Specifically, the tasks might be relying on external files (such as `docker-compose.yml`) without gradle being aware of this, and so when only those files change, gradle doesn't know that the test tasks should be re-run and considers them up-to-date.