Implement Error Prone `ThrowError` to discourage throwing Errors in production code (#957)
Errors are often handled poorly by libraries resulting in unexpected
behavior and resource leaks. It's not obvious that 'catch (Exception e)'
does not catch Error.

This check  is intended to be advisory - it's fine to
`@SuppressWarnings("ThrowError")` in certain cases, but is usually not
recommended unless you are writing a testing library that throws
AssertionError.