NonComparableStreamSort validates direct types
Previously NonComparableStreamSort only failed on types which
would cause a compilation error casting to Comparable. Now we
validate that the stream type is a subtype of Comparable.

This commit reduces the severity to WARNING because it's
possible that the check will flag code that happens to work
today, but we strongly recommend against sorting streams of
a type that is not directly comparable without a custom
comparator becuase it is likely to break later due to lack
of enforcement by the type system.