Allow non-default errorprone fixes to be applied from the command line
You may apply specific error-prone refactors including those which are
not enabled by default by providing a comma delimited list of check
names to the `-PerrorProneApply` option.

```bash
./gradlew compileJava compileTestJava -PerrorProneApply=PreferAssertj
```

In an interal project we have used a similar flag to test refactors
prior to rollout, and I have wished that it was available to help
with Assertj rollout to avoid temporarily modifying gradle
configurations.