Fix PreferSafeLogger edge case that produced non-compiling code
In most cases we wouldn't go this far to ensure green builds, these
are easy enough to manually update. However, it's important that
automation can roll out the new facade without blocking for devs
to make changes, and we can follow-up with a strict check later
on to migrate the rest.