Idea integration results in correct copyright xml files
Previously we had:

```xml
<component name="CopyrightManager">
  <copyright>
    <option name="notice" value="(c) Copyright ${today.year} Palantir
Technologies Inc. All rights reserved.&#10;&#10;Licensed under the
Apache License, Version 2.0 (the &quot;License&quot;);&#10;you may not
use this file except in compliance with the License.&#10;You may obtain
a copy of the License at&#10;&#10;
http://www.apache.org/licenses/LICENSE-2.0&#10;&#10;Unless required by
applicable law or agreed to in writing, software&#10;distributed under
the License is distributed on an &quot;AS IS&quot; BASIS,&#10;WITHOUT
WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.&#10;See
the License for the specific language governing permissions
and&#10;limitations under the License."/>
    <option name="keyword" value="Copyright"/>
    <option name="allowReplaceKeyword" value=""/>
    <option name="myName" value="001_apache-2.0.txt"/>
    <option name="myLocal" value="true"/>
  </copyright>
  <copyright>
    <option name="notice" value="(c) Copyright ${today.year} Palantir
Technologies Inc. All rights reserved.&#10;&#10;Licensed under the
Apache License, Version 2.0 (the &quot;License&quot;);&#10;you may not
use this file except in compliance with the License.&#10;You may obtain
a copy of the License at&#10;&#10;
http://www.apache.org/licenses/LICENSE-2.0&#10;&#10;Unless required by
applicable law or agreed to in writing, software&#10;distributed under
the License is distributed on an &quot;AS IS&quot; BASIS,&#10;WITHOUT
WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.&#10;See
the License for the specific language governing permissions
and&#10;limitations under the License."/>
    <option name="keyword" value="Copyright"/>
    <option name="allowReplaceKeyword" value=""/>
    <option name="myName" value="001_apache-2.0.txt"/>
    <option name="myLocal" value="true"/>
  </copyright>
  <copyright>
    <option name="notice" value="(c) Copyright ${today.year} Palantir
Technologies Inc. All rights reserved.&#10;&#10;Licensed under the
Apache License, Version 2.0 (the &quot;License&quot;);&#10;you may not
use this file except in compliance with the License.&#10;You may obtain
a copy of the License at&#10;&#10;
http://www.apache.org/licenses/LICENSE-2.0&#10;&#10;Unless required by
applicable law or agreed to in writing, software&#10;distributed under
the License is distributed on an &quot;AS IS&quot; BASIS,&#10;WITHOUT
WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.&#10;See
the License for the specific language governing permissions
and&#10;limitations under the License."/>
    <option name="keyword" value="Copyright"/>
    <option name="allowReplaceKeyword" value=""/>
    <option name="myName" value="001_apache-2.0.txt"/>
    <option name="myLocal" value="true"/>
  </copyright>
  <copyright>
    <option name="notice" value="(c) Copyright ${today.year} Palantir
Technologies Inc. All rights reserved.&#10;&#10;Licensed under the
Apache License, Version 2.0 (the &quot;License&quot;);&#10;you may not
use this file except in compliance with the License.&#10;You may obtain
a copy of the License at&#10;&#10;
http://www.apache.org/licenses/LICENSE-2.0&#10;&#10;Unless required by
applicable law or agreed to in writing, software&#10;distributed under
the License is distributed on an &quot;AS IS&quot; BASIS,&#10;WITHOUT
WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.&#10;See
the License for the specific language governing permissions
and&#10;limitations under the License."/>
    <option name="keyword" value="Copyright"/>
    <option name="allowReplaceKeyword" value=""/>
    <option name="myName" value="001_apache-2.0.txt"/>
    <option name="myLocal" value="true"/>
  </copyright>
</component>
```