Results of boolean logic are safe
This resolves an issue in which we considered the result of
a null-check on an exception unsafe, which is not accurate.

The fix is slightly more involved than it would have been due to
the discovery of another bug: auto-boxing and auto-unboxing is
interpreted as method invocations along the lines of
`Boolean.valueOf(bool)` and `Boolean.booleanValue()` respectively.
I added safety-passthrough for such methods.