Stream.collect safety no longer incorrectly includes map collectors
Map collectors, unlike most others, doesn't collect the data from
the stream into an element, rather acts in two steps: A `stream.map`
for each key and value function, as well as a collect of the merged
results. Thus, we cannot make assumptions about the output safety
based on the inputs. In the future, this will be handled using
the same functionality as `stream.map(func)`.