Set the java launcher for Checkstyle tasks, too
The gradle-jdks plugin aims to download JDKs from a source and use them
at runtime. In theory, this means that in order to develop using such a
library, one only needs a JDK installed that can run Gradle (there are
some limitations but this is the theory).

Inside the Checkstyle task is a Java launcher property. When this is
resolved, it will query the Java version service, which will query
the configured JDKs. This means that the baseline-java-versions
configured JDK is not used, and if the gradle-jdks approach is used, the
Checkstyle task will fail.

I _think_ that this change is enough to correctly override the Java
launcher used.

Fixes https://github.com/palantir/gradle-jdks/issues/93