Workaround to IDEA-301084
See https://github.com/JetBrains/intellij-community/pull/2135 for
more details. Even if JetBrains takes this PR, it'll be months
before it's common that all devs have a fixed version (it's common
for devs to be a version or two behind) so it seems sensible to
just workaround in baseline for hte next year or two.