Exclude Spark UnusedStubClass from class uniqueness
Spark UnusedStubClass exists for mechanical reasons and is as described,
never used. With this PR it is skipped, like module-infos.

Potential badness: It's possible that this change would block excavators
on certain projects by changing the set of duplicated classes. I think
it's likely that this would simply require approval from a human,
depending on how the excavator was written.