ensure tasks that use moduleJvmArgs have task inputs set correctly
The BaselineModuleJvmArgs plugin configures several tasks from the
'java' plugin to use values from the 'moduleJvmArgs' extension (e.g. to
update compiler arguments, output values to the jar manifest, etc.).
However, when the value on the extension changes, those tasks are not
re-run.

To fix that, add input properties for each task that uses moduleJvmArgs
so that when the extension value changes, the task will no longer be
up-to-date.

Fixes #2476