Allow leading _ in Java local var names
errorprone: CheckReturnValue requires methods with a return value use them (and @CanIgnoreReturnValue isn't an option).
If you really don't want to use the result, then assign it to a variable: `var unused = ...`

But doing that then results in another errorprone: StrictUnusedVariable: The local variable 'unused' is never read. Intentional occurrences are acknowledged by renaming the unused variable with a leading underscore. '_unused'

Checkstyles localVariableName rejects that (however allows method args to have leading _).

This change allow local vars to have a leading _ to align with errorProne and be consistent with method vars.