cache key for JarClassHasher includes artifact classifier
Previously, we used the `ModuleVersionIdentifier` associated with a
`ResolvedArtifact` as the key for this cache, which contains sha256
hashes of each class file within a jar. Unfortunately this is not
sufficient for distinguishing artifacts with the same maven coordinate,
but different classifiers, which may contain completely different
content within the jars.

This leads to problems when two artifacts from the same maven coordinate
but different classifiers are on the classpath, wherein
`checkClassUniqueness` potentially reports incorrect results as it may
rely on hashes cached from a different artifact altogether (but for the
same `ModuleVersionIdentifier`).